[VAS-720] Reworked AddEditCommissionBundleForm
