+++ src/locale/it.json
@@ -899,6 +899,16 @@
         "from": "Da",
         "to": "A"
       },
+      "addTaxonomies":{
+        "subTitle": "Cerca i servici di incasso sfogliando il catalogo oppure carica una lista di codici.",
+        "manualHelp": "Dubbi? Vai al manuale",
+        "catalogueButton": "Sfoglia il catalogo",
+        "dropFileText": "oppure lascia qui una lista di codici di incasso",
+        "addedServices": "Servizi aggiunti",
+        "catalogueTitle": "Catalogo servizi",
+        "catalogueSubtitle": "Sfoglia il catalogo o cerca per codice di incasso",
+        "filterTitle": "Codice di incasso"
+      },
       "validationMessage": {
         "requiredField": "Campo obbligatorio",
         "dateNotValid": "La data inserita non è valida",
@@ -915,6 +925,7 @@
         "errorMessageAllDataDesc": "Errore durante la ricezione dei dati",
         "errorMessageChannelIdsDataDesc": "Errore durante la ricezione dei canali",
         "errorMessageCreatingBundle": "Errore durante la creazione del pacchetto commissioni",
+        "errorMessageUpdatingBundle": "Errore durante l'aggiornamento del pacchetto commissioni",
         "errorMessageNoBrokerDelegations": "L’intermediario non ha ancora effettuato la registrazione al Nodo dei Pagamenti"
       }
     }
